fix: rename `aea_ledger_hwi` -> `aea_ledger_ethereum_hwi`
