+++ plugins/aea-ledger-ethereum-hwi/LICENSE
