+++ plugins/aea-ledger-ethereum-hwi/MANIFEST.in
@@ -1,4 +1,4 @@
 include README.md LICENSE HISTORY.md
 
-recursive-include aea_ledger_hwi
+recursive-include aea_ledger_ethereum_hwi
 recursive-include tests *
