+++ plugins/aea-ledger-ethereum-hwi/aea_ledger_ethereum_hwi/exceptions.py
