+++ plugins/aea-ledger-ethereum-hwi/aea_ledger_ethereum_hwi/hwi.py
@@ -34,7 +34,7 @@ from aea_ledger_ethereum.ethereum import (
     TESTNET_NAME,
     set_wrapper_for_web3py_session_cache,
 )
-from aea_ledger_hwi.account import HWIAccount
+from aea_ledger_ethereum_hwi.account import HWIAccount
 from eth_account.messages import encode_defunct
 from eth_account.signers.local import LocalAccount
